[bsg] refactor by splitting the recording identification component into smaller parts
