+++ projects/kerttu-global/src/app/+identification/recording-identification/identification-main/identification-view/identification-table/identification-panel/identification-panel.component.html
