+++ projects/kerttu-global/src/app/+identification/recording-identification/identification-main/identification-view/identification-view.component.scss
@@ -1,4 +1,4 @@
-@import 'vars';
+@import '../../../../../../../laji-ui/src/lib/vars';
 
 .top-bar {
   overflow: auto;
