+++ projects/kerttu-global/src/app/+identification/recording-identification/identification-main/identification-view/identification-view.component.ts
@@ -23,29 +23,29 @@ import {
   IGlobalSpeciesWithAnnotation,
   SpeciesAnnotationEnum,
   TaxonTypeEnum
-} from '../../../kerttu-global-shared/models';
+} from '../../../../kerttu-global-shared/models';
 import {
   AudioViewerMode,
   IAudioViewerArea,
   IAudioViewerRectangle,
   ISpectrogramConfig
-} from '../../../../../../laji/src/app/shared-modules/audio-viewer/models';
+} from '../../../../../../../laji/src/app/shared-modules/audio-viewer/models';
 import { map } from 'rxjs/operators';
-import { KerttuGlobalApi } from '../../../kerttu-global-shared/service/kerttu-global-api';
+import { KerttuGlobalApi } from '../../../../kerttu-global-shared/service/kerttu-global-api';
 import { forkJoin, Observable, Subscription } from 'rxjs';
 import { TranslateService } from '@ngx-translate/core';
-import { KerttuGlobalUtil } from '../../../kerttu-global-shared/service/kerttu-global-util.service';
+import { KerttuGlobalUtil } from '../../../../kerttu-global-shared/service/kerttu-global-util.service';
 import { IdentificationTableComponent } from './identification-table/identification-table.component';
-import { defaultSpectrogramConfig } from '../../../../../../laji/src/app/shared-modules/audio-viewer/variables';
+import { defaultSpectrogramConfig } from '../../../../../../../laji/src/app/shared-modules/audio-viewer/variables';
 import {
   defaultAudioSampleRate,
   defaultBatAudioSampleRate,
   defaultInsectAudioSampleRate,
   lowAudioSampleRate
-} from '../../../kerttu-global-shared/variables';
+} from '../../../../kerttu-global-shared/variables';
 import { DOCUMENT } from '@angular/common';
-import { AudioViewerComponent } from '../../../../../../laji/src/app/shared-modules/audio-viewer/audio-viewer/audio-viewer.component';
-import { ModalRef, ModalService } from '../../../../../../laji-ui/src/lib/modal/modal.service';
+import { AudioViewerComponent } from '../../../../../../../laji/src/app/shared-modules/audio-viewer/audio-viewer/audio-viewer.component';
+import { ModalRef, ModalService } from '../../../../../../../laji-ui/src/lib/modal/modal.service';
 
 
 @Component({
