+++ projects/kerttu-global/src/app/+identification/recording-identification/identification-main/identification-view/taxon-select/taxon-select.component.scss
@@ -1,4 +1,4 @@
-@import 'vars';
+@import '../../../../../../../../laji-ui/src/lib/vars';
 
 :host ::ng-deep typeahead-container.dropdown {
   margin-left: 0;
