+++ projects/kerttu-global/src/app/+identification/recording-identification/identification-main/identification-view/taxon-select/taxon-select.component.ts
@@ -1,9 +1,9 @@
 import { catchError, map, switchMap, tap } from 'rxjs/operators';
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, EventEmitter, Input, Output } from '@angular/core';
 import { Observable, of, of as ObservableOf } from 'rxjs';
-import { KerttuGlobalApi } from '../../../../kerttu-global-shared/service/kerttu-global-api';
-import { UserService } from '../../../../../../../laji/src/app/shared/service/user.service';
-import { IGlobalSpecies, IGlobalSpeciesFilters, TaxonTypeEnum } from '../../../../kerttu-global-shared/models';
+import { KerttuGlobalApi } from '../../../../../kerttu-global-shared/service/kerttu-global-api';
+import { UserService } from '../../../../../../../../laji/src/app/shared/service/user.service';
+import { IGlobalSpecies, IGlobalSpeciesFilters, TaxonTypeEnum } from '../../../../../kerttu-global-shared/models';
 import { TranslateService } from '@ngx-translate/core';
 
 interface IGlobalSpeciesWithAutocompleteInfo extends IGlobalSpecies {
