+++ projects/kerttu-global/src/app/+identification/recording-identification/site-selection/site-selection-view/site-selection-map/site-selection-map.component.html
