+++ projects/kerttu-global/src/app/+identification/recording-identification/site-selection/site-selection-view/site-selection-view.component.scss
@@ -1,4 +1,4 @@
-@import 'projects/laji-ui/src/lib/vars';
+@import '../../../../../../../laji-ui/src/lib/vars';
 
 
 @media only screen and (min-width: map-get($map: $breakpoints, $key: "sm")) {
