+++ projects/kerttu-global/src/app/+identification/recording-identification/site-selection/site-selection-view/site-table/site-table.component.html
